requirement.txt
mediapipe==0.9.1
numpy==1.24.2
opencv-python==4.7.0.72
pillow==9.4.0
pyaudio==0.2.11

